feat: consolidate redundant multi-objective optimization files
- Created unified multi-objective optimizer at src/utils/nas_tas/unified_multi_objective.py
- Updated all multi-objective files to use unified implementation:
  - evaluation/multi_objective.py now uses unified optimizer
  - tas_regime/search/multi_objective_search.py now uses unified optimizer
  - hybrid_nas_tas_regime/shared_utils/performance_estimators.py now redirects to unified optimizer
- Eliminated ~2,000+ lines of duplicate multi-objective code
- Maintained backward compatibility for all existing imports
- Consolidated 3 multi-objective implementations into 1 unified implementation

Benefits:
- Single source of truth for multi-objective optimization
- Reduced maintenance overhead
- Consistent behavior across all optimizers
- Improved performance through unified optimization
- Support for both NAS and TAS architectures
